% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_momo.R
\name{launch_momo}
\alias{launch_momo}
\title{Alternative launcher for EuroMoMo code.}
\usage{
launch_momo(MOMOfile, hfile)
}
\arguments{
\item{MOMOfile}{mortality data \code{data.frame}. It must contain columns \code{DoD} (\code{Date} class, date of death), 
\code{DoR} (\code{Date} class, reporting date), and \code{age} (\code{numeric}); other columns can be added 
(see EuroMoMo documentation).}

\item{hfile}{bank holidays \code{data.frame}. It must contain columns \code{date} and \code{closed}.
The master start button for running all of the MoMo code. \code{\link{SetOpts}} must be run first.}
}
\description{
It is an alternative launcher for the EuroMoMo code that accepts data in R format. It avoids the need to create
intermediate text/Stata files if data has been captured in R previously.
}
\examples{
SetOpts(
  DoA=as.Date("2013-12-31"),
  DoPR=as.Date("2008-1-1"),
  WStart=1,
  WEnd=52,
  country = "Denmark",
  source = "SSI",
  MFILE = "DoD_DoR.txt",
  HFILE = "holidays.txt",
  INPUTDIR = system.file("testdata",package="momo"),
  WDIR = tempdir(),
  back = 3,
  WWW = 290,
  Ysum = 2013,
  Wsum = 40,
  USEglm2 = TRUE,
  useAUTOMN = TRUE,
  datesISO = FALSE,
  plotGraphs = FALSE
)

momo_file_name <- system.file("testdata", "DoD_DoR.txt", package = "momo")

MOMOfile <- read.csv(momo_file_name, sep = ";", stringsAsFactors = FALSE)
MOMOfile$DoD <- as.Date(MOMOfile$DoD)
MOMOfile$DoR <- as.Date(MOMOfile$DoR)

holidays_file_name <- system.file("testdata", "holidays.txt", package = "momo")
hfile <- read.csv(holidays_file_name, sep = ";", stringsAsFactors = FALSE)
hfile$date <- as.Date(hfile$date)

launch_momo(MOMOfile, hfile)
}
