% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_MOMOmaster_set_and_run.R
\name{SetOpts}
\alias{SetOpts}
\title{Placeholder list for all the following options}
\usage{
SetOpts(DoA = as.Date("2015-8-10"), DoPR = as.Date("2008-1-1"),
  WStart = 1, WEnd = 52, country = "Greece", source = "UoP",
  MFILE = "greece_m.dta", HFILE = "holidayfilegreece.dta",
  INPUTDIR = "./input", WDIR = "./output", back = 3, WWW = 290,
  Ysum = 2009, Wsum = 34, USEglm2 = TRUE, useAUTOMN = TRUE,
  datesISO = FALSE, plotGraphs = TRUE)
}
\description{
Placeholder list for all the following options
}
